refactor: removed duplicate LoadingMask file
